base/requirement.txt
amqp==5.3.1
asgiref==3.10.0
billiard==4.2.2
bleach==6.3.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
dj-stripe==2.10.3
Django==5.2.7
django-admin-interface==0.30.1
django-allauth==65.13.0
django-axes==8.0.0
django-bootstrap==0.2.4
django-ckeditor==6.7.3
django-colorfield==0.14.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-extensions==4.1
django-guardian==3.2.0
django-import-export==4.3.13
django-jet==1.0.8
django-js-asset==3.1.2
django-paypal==2.1
django-restframework==0.0.1
django-storages==1.14.6
django-summernote==0.8.20.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==13.1.1
tablib==3.9.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
